Parallax, lighting updates, and health/mana
Level is now has sidescrolling functionality with variable width and
height. Lighting coloration updated, and font and art assets updated
(including multiple layers of background and foreground parallaxes).
Ground enemies now deal damage, and health and mana are tracked and used
for win/lose conditions and spell casting, and are displayed on screen
in Diablo-esque orbs.