Changes to enemy spawning/AI and progression
Enemy spawning now works based on a 1 second looping timer, and is now
limited by the number of enemies on screen instead of the number in the
world. Bats now shoot poison bolt projectiles, and fly higher above the
player.

Groundwork laid out for different states of game progression, which will
later allow for repairing the wizard's tower as waystones are reaches.
Waystones now increment the progression variable, and give a tooltip
when walked over indicating they can be used for warping.

Generally cleaned up a few variables; pausing is now a game state, which
has taken the place of the unused casting game state. This has removed
the paused variable, and also fixes some random rare bugs which have to
do with incorrectly checking if the game was in a state of being played,
since paused had to be checked separately from game states (and wasn't
always being checked).