Wizard tower, player sprite sheet, changes to game states
Added wizard tower that spawns at map center with attached light
sources. Player now animates from a sprite sheet, including a transition
to an idle pose when not moving. Removed high score screen and game
state as they are a relic of the engine and aren't used for this game.
Added offset property to LightSource class to allow light sources to be
offset from their root object. Used by the tower to place light sources
at varying points around it.